Fixed issue with integer overflow when calculating correlation coefficient.

+++ src/correlation_coefficient.c
@@ -10,11 +10,11 @@
 
 double corrcoef(int n, int x[], int y[]) {
 
-    int sum_x = 0;
-    int sum_y = 0;
-    int sum_xy = 0;
-    int squareSum_x = 0;
-    int squareSum_y = 0;
+    long sum_x = 0;
+    long sum_y = 0;
+    long sum_xy = 0;
+    long squareSum_x = 0;
+    long squareSum_y = 0;
     
     // calculate sums
     for (int i = 0; i < n; i++)
@@ -28,7 +28,7 @@ double corrcoef(int n, int x[], int y[]) {
  
     // calculate correlation
     double correlation = (double)(n * sum_xy - sum_x * sum_y) 
-        / sqrt((n * squareSum_x - sum_x * sum_x)
+        / (double)sqrt((n * squareSum_x - sum_x * sum_x)
         * (n * squareSum_y - sum_y * sum_y));
  
     return correlation;
